% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest.R
\name{mx_filter_studies}
\alias{mx_filter_studies}
\title{Filter studies below a QC threshold}
\usage{
mx_filter_studies(studies, qc_threshold = 7)
}
\arguments{
\item{studies}{A named list of \code{\linkS4class{metaXpressStudy}} objects.}

\item{qc_threshold}{Numeric scalar. Minimum acceptable QC score (inclusive).
Default: \code{7}.}
}
\value{
A filtered list of \code{\linkS4class{metaXpressStudy}} objects.
}
\description{
Removes studies whose \code{qc_score} is below the specified threshold.
Studies with \code{NA} QC scores (i.e., \code{\link{mx_qc_study}} has not
been run) are always removed with a warning.
}
\examples{
\dontrun{
  studies <- mx_fetch_geo(c("GSE12345", "GSE67890"))
  studies <- mx_filter_studies(studies, qc_threshold = 7)
}

}
\seealso{
\code{\link{mx_qc_study}}
}
