% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_visualize.R
\name{mx_heterogeneity_plot}
\alias{mx_heterogeneity_plot}
\title{I-squared distribution plot}
\usage{
mx_heterogeneity_plot(meta_result)
}
\arguments{
\item{meta_result}{A \code{\linkS4class{metaXpressResult}} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Displays the distribution of I-squared heterogeneity values across all
genes in the meta-analysis, with reference lines at 25% and 75%.
}
\examples{
\dontrun{
  mx_heterogeneity_plot(meta_result)
}

}
\seealso{
\code{\link{mx_heterogeneity}}, \code{\link{mx_forest}}
}
