% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_harmonize.R
\name{mx_reannotate}
\alias{mx_reannotate}
\title{Reannotate gene IDs to a common namespace}
\usage{
mx_reannotate(
  studies,
  org = "Homo sapiens",
  target_id = c("SYMBOL", "ENSEMBL", "ENTREZID")
)
}
\arguments{
\item{studies}{A named list of \code{\linkS4class{metaXpressStudy}} objects.}

\item{org}{Character scalar. Target organism. Currently only
\code{"Homo sapiens"} is supported. Default: \code{"Homo sapiens"}.}

\item{target_id}{Character scalar. Target gene identifier type.
One of \code{"SYMBOL"} (default), \code{"ENSEMBL"}, or
\code{"ENTREZID"}.}
}
\value{
The input list of studies with \code{counts} rownames translated
  to the target namespace.
}
\description{
Converts gene identifiers in one or more studies to a unified namespace
(Ensembl gene ID, gene symbol, or Entrez ID) using \pkg{AnnotationDbi}.
Genes that cannot be mapped are dropped with a warning.
}
\examples{
\dontrun{
  studies <- mx_reannotate(studies, org = "Homo sapiens",
                            target_id = "SYMBOL")
}

}
\references{
Stokholm, Rabaglino & Kadarmideen (2024) Current Protocols.
\doi{10.1002/cpz1.70046}
}
\seealso{
\code{\link{mx_normalize}}, \code{\link{mx_align_genes}}
}
