% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_visualize.R
\name{mx_study_overview}
\alias{mx_study_overview}
\title{Study characteristics overview plot}
\usage{
mx_study_overview(studies)
}
\arguments{
\item{studies}{A named list of \code{\linkS4class{metaXpressStudy}} objects.}
}
\value{
A \code{patchwork} object containing multiple \code{ggplot2} plots.
}
\description{
Creates a multi-panel summary of study-level QC metrics including
QC scores, sample sizes, and sequencing depths.
}
\examples{
\dontrun{
  mx_study_overview(studies)
}

}
