% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{show,metaXpressResult-method}
\alias{show,metaXpressResult-method}
\title{Display a metaXpressResult object}
\usage{
\S4method{show}{metaXpressResult}(object)
}
\arguments{
\item{object}{A \code{metaXpressResult} object.}
}
\value{
Invisibly returns \code{object}. Called for its side-effect of
  printing a summary to the console.
}
\description{
Display a metaXpressResult object
}
