% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{show,metaXpressStudy-method}
\alias{show,metaXpressStudy-method}
\title{Display a metaXpressStudy object}
\usage{
\S4method{show}{metaXpressStudy}(object)
}
\arguments{
\item{object}{A \code{metaXpressStudy} object.}
}
\value{
Invisibly returns \code{object}. Called for its side-effect of
  printing a summary to the console.
}
\description{
Display a metaXpressStudy object
}
