% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest.R
\name{mx_cluster_samples}
\alias{mx_cluster_samples}
\title{Automatically cluster samples by metadata}
\usage{
mx_cluster_samples(study)
}
\arguments{
\item{study}{A \code{\linkS4class{metaXpressStudy}} object.}
}
\value{
The input \code{study} with the \code{condition} column updated
  in \code{metadata}.
}
\description{
Uses metadata fields to assign samples to case/control groups, mimicking
the sampleclusteR approach for automated group assignment from GEO
metadata.
}
\examples{
\dontrun{
  study <- mx_cluster_samples(study)
}

}
\references{
Coke, T., Niranjan, M. & Ewing, R.M. (2025) sampleclusteR: automated
case/control group assignment from GEO metadata. \emph{bioRxiv}.
\doi{10.1101/2025.04.10.648129}
}
